// src/types/resources/Practitioner.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { THumanName } from '../partials/HumanName';
import { TContactPoint } from '../partials/ContactPoint';
import { TAddress } from '../partials/Address';
import { TDate } from '../simpleTypes/Date';
import { TAttachment } from '../partials/Attachment';
import { TPractitionerQualification } from '../partials/PractitionerQualification';
import { TCodeableConcept } from '../partials/CodeableConcept';

export type TPractitioner = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  active?: boolean;
  name?: THumanName[];
  telecom?: TContactPoint[];
  address?: TAddress[];
  gender?: string;
  birthDate?: TDate;
  photo?: TAttachment[];
  qualification?: TPractitionerQualification[];
  communication?: TCodeableConcept[];
};